# Kernel agnostic mcmc from deformable beta splatting