@RongLiu-Leo 2025-05-13
Another practice is that only do the opacity regularization during densification to make distribution preserved. Otherwise the scene looks too transparent at the end. Cancelling this regularization won't hurt the metrics but produces more visual appealing results.

Change 
```
if cfg.opacity_reg > 0.0:
        loss = (
            loss
            + cfg.opacity_reg
            * torch.abs(torch.sigmoid(self.splats["opacities"])).mean()
        )
```
to something like
```
if cfg.opacity_reg > 0.0 and iteration < cfg.strategy.refine_stop_iter:
        loss = (
            loss
            + cfg.opacity_reg
            * torch.abs(torch.sigmoid(self.splats["opacities"])).mean()
        )
```