@liruilong940607 2025-05-14
I like that it simplify the MCMC and still keep the same metrics!!!

But hurt runtime is less ideal. If we can get at least same runtime I would be happy to merge it